Bug Fixes
playFMNewGame() has been changed back to private and instead the main() calls a new game to be played, to which the game menu will call SolitaireFM.main(). This goes for all other game types we implement.
Bug Fixes: game logic is now roughly 90% where we want it, although there is some erroneous movement that does not conform to validPlayStackMove().