Local User File Creation
Files will now be created in ".solitairegame" in the the 'User' folder that windows uses.
Analysis will create two files, one that is on a single line and another that does a pretty print of XML.
On running of the game, it will no longer print the cards suit and value during initial card dealing. This issue was due the way previous code handled this. I created a constructor and toString within the Enums.