Product Backlog Order: 19-22, 24-30
Changes listed in the Product/Sprint backlog, however, 23 hasn't been implemented (the additional XML structure for game types).
The setup of the program has changed--the audio files are not under 'global/sound'. And the Card, CardStack, and FinalStack are not consolidated for all game types. This will help facilitate the implementation by other Teams.